requirements.txt
Django>==4.1.7
Pillow>==9.4.0
psycopg2>==2.9.5
django-filter>==23.1
pip>==22.3.1
djangorestframework>==3.14.0